api.legacy: move transplants to support
